feat(api): migrate to background batch processing for news summaries
Introduces a dedicated Cloudflare Worker for scheduled RSS fetching and AI summarization, reducing main API latency by serving pre-processed content from KV.
