+++ functions/api/news.ts
@@ -1,131 +1,11 @@
-import glossary from "./glossary.json";
-import { extractFullContent, SOURCES, getThumbnail } from "./extractors";
-import { cleanHtml, smartTruncate } from "./utils";
-
-interface Ai {
-  run(model: string, input: Record<string, unknown>): Promise<{ response?: string }>;
-}
-
-interface Env {
-  NEWS_TRANSLATIONS: KVNamespace;
-  AI: Ai;
-}
-
-async function generateId(url: string): Promise<string> {
-  const msgUint8 = new TextEncoder().encode(url);
-  const hashBuffer = await crypto.subtle.digest('SHA-256', msgUint8);
-  const hashArray = Array.from(new Uint8Array(hashBuffer));
-  return hashArray.map(b => b.toString(16).padStart(2, '0')).join('');
-}
-
-async function generateFullSummary(ai: Ai, text: string): Promise<string | null> {
-  if (!text) return null;
-  const truncatedText = smartTruncate(text, 3000);
-  console.log("--- AI INPUT START ---");
-  console.log(truncatedText);
-  console.log("--- AI INPUT END ---");
-  try {
-    const prompt = `### 役割
-あなたはプロの日本人ニュースライターです。提供された英語のニュース情報を基に、オーストラリア在住の日本人向けに、日本の大手ニュースサイトに掲載されるような自然な日本語記事を執筆してください。
-
-### 執筆ルール
-1. 段落構成(3段落のみ):
-   - 第1段落(リード):最重要情報を150文字で簡潔にまとめる。
-   - 第2段落:背景・状況・関係者の短いコメントを250文字で説明する。
-   - 第3段落:影響・今後の見通しを250文字で述べる。
-
-2. 禁止事項:
-   - 段落番号やラベルを含めない。
-   - 見出し・箇条書き・注釈を含めない。
-   - 同じ内容を繰り返さない。
-   - 引用は1文以内にする。
-
-3. 出力形式:
-   - 全体の文字数は **650文字** に収める。
-   - 文体は簡潔で事実ベース。
-   - 文章は途中で切らず、最後まで書き切る。
-
-### 英語ニュース本文
-${truncatedText}
-
-### 日本語記事執筆結果（本文のみを出力）:`;
-
-    const response = await ai.run("@cf/meta/llama-3.1-8b-instruct", {
-      prompt,
-      max_tokens: 800
-    }, {
-      gateway: {
-        id: "default",
-        skipCache: false,
-        cacheTtl: 3600
-      }
-    });
-
-    return response.response?.trim() || null;
-  } catch (e) {
-    console.error("Summary generation error:", e);
-    return null;
-  }
-}
-
-function applyGlossary(text: string): string {
-  let processed = text;
-  for (const [short, full] of Object.entries(glossary)) {
-    const regex = new RegExp(`\\b${short}\\b`, 'g');
-    processed = processed.replace(regex, full);
-  }
-  return processed;
-}
-
-async function translateText(ai: Ai, text: string): Promise<string | null> {
-  if (!text) return null;
-  try {
-    const expandedText = applyGlossary(text);
-    const prompt = `### 役割
-あなたはプロのニュースエディターです。提供された英語のニュースタイトルを基に、日本のニュースサイト（Yahoo!ニュース等）で目を引くような、簡潔でインパクトのある日本語の見出しを作成してください。
-
-### 執筆ルール
-- 30〜40文字程度の「見出し」として作成してください。
-- 翻訳調を避け、ニュースらしい体言止めや力強い表現を用いてください。
-- 説明的な文章（〜ました、〜困っている等）ではなく、事実や核心を突く表現にしてください。
-- 見出しのみを出力し、注釈やメタ情報は一切含めないでください。
-
-### 対象タイトル
-${expandedText}
-
-### 日本語見出し（見出しのみを出力）:`;
-
-    const response = await ai.run("@cf/meta/llama-3.1-8b-instruct", {
-      prompt,
-      max_tokens: 300
-    }, {
-      gateway: {
-        id: "default",
-        skipCache: false,
-        cacheTtl: 3600
-      }
-    });
-    return response.response?.trim() || null;
-  } catch (e) {
-    console.error(`Translation error:`, e);
-    return null;
-  }
-}
-
-function extractTagContent(itemXml: string, tagName: string): string {
-  const regex = new RegExp(`<${tagName}[^>]*>([\\s\\S]*?)<\\/${tagName}>`, 'i');
-  const match = itemXml.match(regex);
-  if (match) return match[1].replace(/<!\[CDATA\[([\s\S]*?)\]\]>/g, '$1').trim();
-  return "";
-}
-
-function extractAllCategories(itemXml: string): string[] {
-  const categories: string[] = [];
-  const regex = /<category[^>]*>([\s\S]*?)<\/category>/gi;
-  let match;
-  while ((match = regex.exec(itemXml)) !== null) categories.push(cleanHtml(match[1]));
-  return categories;
-}
+import { SOURCES, getThumbnail } from "./extractors";
+import { 
+  Env, 
+  generateId, 
+  extractTagContent, 
+  extractAllCategories 
+} from "./shared";
+import { cleanHtml } from "./utils";
 
 export const onRequest: PagesFunction<Env> = async (context) => {
   const { env, request } = context;
@@ -146,6 +26,7 @@ export const onRequest: PagesFunction<Env> = async (context) => {
     const limit = parseInt(url.searchParams.get('limit') || '5');
     const before = parseInt(url.searchParams.get('before') || Date.now().toString());
 
+    // 1. Fetch RSS from all sources
     const rssResponses = await Promise.all(SOURCES.map(s => fetch(s.url, { headers: { "User-Agent": "Mozilla/5.0" } })));
     const allItemsXml: string[] = [];
     for (const res of rssResponses) {
@@ -155,42 +36,33 @@ export const onRequest: PagesFunction<Env> = async (context) => {
       allItemsXml.push(...items);
     }
 
+    // 2. Parse items and generate IDs
     const parsedItems = await Promise.all(allItemsXml.map(async itemXml => {
-      const title = cleanHtml(extractTagContent(itemXml, "title"));
       const link = extractTagContent(itemXml, "link");
       const id = await generateId(link);
-      const descriptionRaw = extractTagContent(itemXml, "description");
       const pubDate = extractTagContent(itemXml, "pubDate");
-      const category = extractAllCategories(itemXml)[0] || "News";
-      const thumbnail = getThumbnail(itemXml, link);
-      const firstLine = cleanHtml(descriptionRaw).substring(0, 300);
-
-      return { id, title, link, firstLine, pubDate: new Date(pubDate).getTime(), displayDate: pubDate, category, thumbnail };
+      return { id, link, pubDate: new Date(pubDate).getTime() };
     }));
 
+    // 3. Filter and Sort
     const uniqueItems = Array.from(new Map(parsedItems.map(item => [item.link, item])).values())
       .sort((a, b) => b.pubDate - a.pubDate)
-      .filter(item => item.pubDate < before)
-      .slice(0, limit);
+      .filter(item => item.pubDate < before);
 
-    const results = await Promise.all(uniqueItems.map(async (item) => {
-      const cacheKey = `ja:id:${item.id}`;
-      const cached = await env.NEWS_TRANSLATIONS.get(cacheKey);
+    // 4. Fetch from KV (Only return what has been batch-processed)
+    const results: any[] = [];
+    for (const item of uniqueItems) {
+      if (results.length >= limit) break;
       
-      if (cached) return JSON.parse(cached);
-
-      // New translation & summary
-      const title_ja = await translateText(env.AI, item.title) || item.title;
-      const fullText = await extractFullContent(item.link);
-      const bodyJa = await generateFullSummary(env.AI, fullText) || "要約を生成できませんでした。";
-
-      const newsItem = { id: item.id, title_ja, bodyJa, link: item.link, thumbnail: item.thumbnail, category: item.category, pubDate: item.displayDate };
-      await env.NEWS_TRANSLATIONS.put(cacheKey, JSON.stringify(newsItem), { expirationTtl: 259200 });
-      return newsItem;
-    }));
+      const cached = await env.NEWS_TRANSLATIONS.get(`ja:id:${item.id}`);
+      if (cached) {
+        results.push(JSON.parse(cached));
+      }
+    }
 
     return new Response(JSON.stringify(results), { headers: { "Content-Type": "application/json; charset=utf-8" } });
-  } catch {
-    return new Response(JSON.stringify({ error: "Failed" }), { status: 500 });
+  } catch (error) {
+    console.error("API Error:", error);
+    return new Response(JSON.stringify({ error: "Failed to fetch news list" }), { status: 500 });
   }
 };
